requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
PyMySQL
python-dotenv
Flask-WTF
Flask-Login
email-validator
Pillow
web3
python-slugify
WTForms-SQLAlchemy
flask-limiter
Flask-Caching
APScheduler
python-dateutil
pytz
bleach
